chore: update .gitignore and remove build artifacts
- Add comprehensive Android .gitignore rules
- Remove .gradle, .idea, build folders from git tracking
- Remove compiled files and build artifacts
- Keep only source code and necessary config files

+++ app/src/test/java/com/mlk/taskmanager/data/repository/TaskRepositoryTest.kt
@@ -97,7 +97,7 @@ class TaskRepositoryTest {
         taskRepository.insertTask(task)
         
         // Then
-        verify(projectRepository).incrementTaskCount(projectId.toInt())
+        verify(projectRepository).incrementTaskCount(projectId)
     }
     
     @Test
@@ -158,8 +158,8 @@ class TaskRepositoryTest {
         taskRepository.updateTask(updatedTask)
         
         // Then
-        verify(projectRepository).decrementTaskCount(oldProjectId.toInt())
-        verify(projectRepository).incrementTaskCount(newProjectId.toInt())
+        verify(projectRepository).decrementTaskCount(oldProjectId)
+        verify(projectRepository).incrementTaskCount(newProjectId)
     }
     
     @Test
@@ -197,7 +197,7 @@ class TaskRepositoryTest {
         taskRepository.deleteTask(task)
         
         // Then
-        verify(projectRepository).decrementTaskCount(projectId.toInt())
+        verify(projectRepository).decrementTaskCount(projectId)
     }
     
     @Test
